Correct the type of container index and size from int to size_t

+++ node-addon-api/log_container.h
@@ -51,7 +51,7 @@ struct MatchPattern{
 
 struct MatchRule
 {
-   int patternIndex;
+   size_t patternIndex;
    bool isExclude;
 };
